Moved menu to its own folder. Added tab interactions
